rename analysis module to remix-analyzer
