+++ remix-analyzer/index.js
@@ -1,4 +1,4 @@
-var CodeAnalysis = require('./src/analysis/staticAnalysisRunner')
+var CodeAnalysis = require('./src/solidity-analyzer')
 
 module.exports = {
   CodeAnalysis: CodeAnalysis
