+++ remix-analyzer/src/solidity-analyzer/index.js
