+++ remix-analyzer/src/solidity-analyzer/modules/assignAndCompare.js
