+++ remix-analyzer/src/solidity-analyzer/modules/categories.js
