+++ remix-analyzer/src/solidity-analyzer/modules/checksEffectsInteraction.js
