+++ remix-analyzer/src/solidity-analyzer/modules/constantFunctions.js
