+++ remix-analyzer/src/solidity-analyzer/modules/deleteDynamicArrays.js
