+++ remix-analyzer/src/solidity-analyzer/modules/gasCosts.js
