+++ remix-analyzer/src/solidity-analyzer/modules/guardConditions.js
