+++ remix-analyzer/src/solidity-analyzer/modules/inlineAssembly.js
