+++ remix-analyzer/src/solidity-analyzer/modules/lowLevelCalls.js
