+++ remix-analyzer/src/solidity-analyzer/modules/noReturn.js
