+++ remix-analyzer/src/solidity-analyzer/modules/selfdestruct.js
