+++ remix-analyzer/src/solidity-analyzer/modules/similarVariableNames.js
