+++ remix-analyzer/src/solidity-analyzer/modules/staticAnalysisCommon.js
