+++ remix-analyzer/src/solidity-analyzer/modules/thisLocal.js
